NODE-412: Code review: Extract InitialSynchronization implementations into separate classes
